Move diagnostic APIs from ConnectivityMgr to DiagnosticDataProvider
