+++ src/include/platform/DiagnosticDataProvider.h
@@ -33,10 +33,20 @@ namespace DeviceLayer {
 // static for the duration of the thread.
 static constexpr size_t kMaxThreadNameLength = 32;
 
+// 48-bit IEEE MAC Address or a 64-bit IEEE MAC Address (e.g. EUI-64).
+constexpr size_t kMaxHardwareAddrSize = 8;
+
 struct ThreadMetrics : public app::Clusters::SoftwareDiagnostics::Structs::ThreadMetrics::Type
 {
     char NameBuf[kMaxThreadNameLength + 1];
     ThreadMetrics * Next; /* Pointer to the next structure.  */
+};
+
+struct NetworkInterface : public app::Clusters::GeneralDiagnostics::Structs::NetworkInterfaceType::Type
+{
+    char Name[Inet::InterfaceId::kMaxIfNameLength];
+    uint8_t MacAddress[kMaxHardwareAddrSize];
+    NetworkInterface * Next; /* Pointer to the next structure.  */
 };
 
 /**
@@ -92,6 +102,14 @@ public:
     virtual CHIP_ERROR GetActiveRadioFaults(GeneralFaults<kMaxRadioFaults> & radioFaults);
     virtual CHIP_ERROR GetActiveNetworkFaults(GeneralFaults<kMaxNetworkFaults> & networkFaults);
 
+    /*
+     * Get the linked list of network interfaces of the current plaform. After usage, each caller of GetNetworkInterfaces
+     * needs to release the network interface list it gets via ReleaseNetworkInterfaces.
+     *
+     */
+    virtual CHIP_ERROR GetNetworkInterfaces(NetworkInterface ** netifpp);
+    virtual void ReleaseNetworkInterfaces(NetworkInterface * netifp);
+
     /**
      * Software Diagnostics methods.
      */
@@ -106,6 +124,38 @@ public:
      */
     virtual CHIP_ERROR GetThreadMetrics(ThreadMetrics ** threadMetricsOut);
     virtual void ReleaseThreadMetrics(ThreadMetrics * threadMetrics);
+
+    /**
+     * Ethernet network diagnostics methods
+     */
+    virtual CHIP_ERROR GetEthPHYRate(uint8_t & pHYRate);
+    virtual CHIP_ERROR GetEthFullDuplex(bool & fullDuplex);
+    virtual CHIP_ERROR GetEthCarrierDetect(bool & carrierDetect);
+    virtual CHIP_ERROR GetEthTimeSinceReset(uint64_t & timeSinceReset);
+    virtual CHIP_ERROR GetEthPacketRxCount(uint64_t & packetRxCount);
+    virtual CHIP_ERROR GetEthPacketTxCount(uint64_t & packetTxCount);
+    virtual CHIP_ERROR GetEthTxErrCount(uint64_t & txErrCount);
+    virtual CHIP_ERROR GetEthCollisionCount(uint64_t & collisionCount);
+    virtual CHIP_ERROR GetEthOverrunCount(uint64_t & overrunCount);
+    virtual CHIP_ERROR ResetEthNetworkDiagnosticsCounts();
+
+    /**
+     * WiFi network diagnostics methods
+     */
+    virtual CHIP_ERROR GetWiFiBssId(ByteSpan & value);
+    virtual CHIP_ERROR GetWiFiSecurityType(uint8_t & securityType);
+    virtual CHIP_ERROR GetWiFiVersion(uint8_t & wiFiVersion);
+    virtual CHIP_ERROR GetWiFiChannelNumber(uint16_t & channelNumber);
+    virtual CHIP_ERROR GetWiFiRssi(int8_t & rssi);
+    virtual CHIP_ERROR GetWiFiBeaconLostCount(uint32_t & beaconLostCount);
+    virtual CHIP_ERROR GetWiFiBeaconRxCount(uint32_t & beaconRxCount);
+    virtual CHIP_ERROR GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount);
+    virtual CHIP_ERROR GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount);
+    virtual CHIP_ERROR GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount);
+    virtual CHIP_ERROR GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount);
+    virtual CHIP_ERROR GetWiFiCurrentMaxRate(uint64_t & currentMaxRate);
+    virtual CHIP_ERROR GetWiFiOverrunCount(uint64_t & overrunCount);
+    virtual CHIP_ERROR ResetWiFiNetworkDiagnosticsCounts();
 
 protected:
     // Construction/destruction limited to subclasses.
@@ -189,6 +239,133 @@ inline CHIP_ERROR DiagnosticDataProvider::GetActiveRadioFaults(GeneralFaults<kMa
 }
 
 inline CHIP_ERROR DiagnosticDataProvider::GetActiveNetworkFaults(GeneralFaults<kMaxNetworkFaults> & networkFaults)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetNetworkInterfaces(NetworkInterface ** netifpp)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline void DiagnosticDataProvider::ReleaseNetworkInterfaces(NetworkInterface * netifp) {}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthPHYRate(uint8_t & pHYRate)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthFullDuplex(bool & fullDuplex)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthCarrierDetect(bool & carrierDetect)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthTimeSinceReset(uint64_t & timeSinceReset)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthPacketRxCount(uint64_t & packetRxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthPacketTxCount(uint64_t & packetTxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthTxErrCount(uint64_t & txErrCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthCollisionCount(uint64_t & collisionCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetEthOverrunCount(uint64_t & overrunCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::ResetEthNetworkDiagnosticsCounts()
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiBssId(ByteSpan & value)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiSecurityType(uint8_t & securityType)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiVersion(uint8_t & wiFiVersion)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiChannelNumber(uint16_t & channelNumber)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiRssi(int8_t & rssi)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiBeaconLostCount(uint32_t & beaconLostCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiBeaconRxCount(uint32_t & beaconRxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiCurrentMaxRate(uint64_t & currentMaxRate)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::GetWiFiOverrunCount(uint64_t & overrunCount)
+{
+    return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
+}
+
+inline CHIP_ERROR DiagnosticDataProvider::ResetWiFiNetworkDiagnosticsCounts()
 {
     return CHIP_ERROR_UNSUPPORTED_CHIP_FEATURE;
 }
