+++ src/platform/ESP32/ConnectivityManagerImpl.cpp
@@ -71,62 +71,5 @@ void ConnectivityManagerImpl::_OnPlatformEvent(const ChipDeviceEvent * event)
 #endif
 }
 
-static InterfaceType GetInterfaceType(const char * if_desc)
-{
-    if (strncmp(if_desc, "ap", strnlen(if_desc, 2)) == 0 || strncmp(if_desc, "sta", strnlen(if_desc, 3)) == 0)
-        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_WI_FI;
-    if (strncmp(if_desc, "openthread", strnlen(if_desc, 10)) == 0)
-        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_THREAD;
-    if (strncmp(if_desc, "eth", strnlen(if_desc, 3)) == 0)
-        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_ETHERNET;
-    return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_UNSPECIFIED;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetNetworkInterfaces(NetworkInterface ** netifpp)
-{
-    esp_netif_t * netif     = esp_netif_next(NULL);
-    NetworkInterface * head = NULL;
-    if (netif == NULL)
-    {
-        ChipLogError(DeviceLayer, "Failed to get network interfaces");
-    }
-    else
-    {
-        for (esp_netif_t * ifa = netif; ifa != NULL; ifa = esp_netif_next(ifa))
-        {
-            NetworkInterface * ifp = new NetworkInterface();
-            strncpy(ifp->Name, esp_netif_get_ifkey(ifa), Inet::InterfaceId::kMaxIfNameLength);
-            ifp->Name[Inet::InterfaceId::kMaxIfNameLength - 1] = '\0';
-            ifp->name                                          = CharSpan(ifp->Name, strlen(ifp->Name));
-            ifp->fabricConnected                               = true;
-            ifp->type                                          = GetInterfaceType(esp_netif_get_desc(ifa));
-            ifp->offPremiseServicesReachableIPv4               = false;
-            ifp->offPremiseServicesReachableIPv6               = false;
-            if (esp_netif_get_mac(ifa, ifp->MacAddress) != ESP_OK)
-            {
-                ChipLogError(DeviceLayer, "Failed to get network hardware address");
-            }
-            else
-            {
-                ifp->hardwareAddress = ByteSpan(ifp->MacAddress, 6);
-            }
-            ifp->Next = head;
-            head      = ifp;
-        }
-    }
-    *netifpp = head;
-    return CHIP_NO_ERROR;
-}
-
-void ConnectivityManagerImpl::_ReleaseNetworkInterfaces(NetworkInterface * netifp)
-{
-    while (netifp)
-    {
-        NetworkInterface * del = netifp;
-        netifp                 = netifp->Next;
-        delete del;
-    }
-}
-
 } // namespace DeviceLayer
 } // namespace chip
