+++ src/platform/ESP32/ConnectivityManagerImpl_WiFi.cpp
@@ -994,154 +994,6 @@ void ConnectivityManagerImpl::OnIPv6AddressAvailable(const ip_event_got_ip6_t & 
     PlatformMgr().PostEventOrDie(&event);
 }
 
-uint8_t MapAuthModeToSecurityType(wifi_auth_mode_t authmode)
-{
-    switch (authmode)
-    {
-    case WIFI_AUTH_OPEN:
-        return 1;
-    case WIFI_AUTH_WEP:
-        return 2;
-    case WIFI_AUTH_WPA_PSK:
-        return 3;
-    case WIFI_AUTH_WPA2_PSK:
-        return 4;
-    case WIFI_AUTH_WPA3_PSK:
-        return 5;
-    default:
-        return 0;
-    }
-}
-
-uint8_t GetWiFiVersionFromAPRecord(wifi_ap_record_t ap_info)
-{
-    if (ap_info.phy_11n)
-        return 3;
-    else if (ap_info.phy_11g)
-        return 2;
-    else if (ap_info.phy_11b)
-        return 1;
-    else
-        return 0;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiBssId(ByteSpan & BssId)
-{
-    wifi_ap_record_t ap_info;
-    esp_err_t err;
-
-    err = esp_wifi_sta_get_ap_info(&ap_info);
-    if (err == ESP_OK)
-    {
-        memcpy(mWiFiMacAddress, ap_info.bssid, 6);
-    }
-    BssId = ByteSpan(mWiFiMacAddress, 6);
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiSecurityType(uint8_t & securityType)
-{
-    securityType = 0;
-    wifi_ap_record_t ap_info;
-    esp_err_t err;
-
-    err = esp_wifi_sta_get_ap_info(&ap_info);
-    if (err == ESP_OK)
-    {
-        securityType = MapAuthModeToSecurityType(ap_info.authmode);
-    }
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiVersion(uint8_t & wifiVersion)
-{
-    wifiVersion = 0;
-    wifi_ap_record_t ap_info;
-    esp_err_t err;
-    err = esp_wifi_sta_get_ap_info(&ap_info);
-    if (err == ESP_OK)
-    {
-        wifiVersion = GetWiFiVersionFromAPRecord(ap_info);
-    }
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiChannelNumber(uint16_t & channelNumber)
-{
-    channelNumber = 0;
-    wifi_ap_record_t ap_info;
-    esp_err_t err;
-
-    err = esp_wifi_sta_get_ap_info(&ap_info);
-    if (err == ESP_OK)
-    {
-        channelNumber = ap_info.primary;
-    }
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiRssi(int8_t & rssi)
-{
-    rssi = 0;
-    wifi_ap_record_t ap_info;
-    esp_err_t err;
-
-    err = esp_wifi_sta_get_ap_info(&ap_info);
-
-    if (err == ESP_OK)
-    {
-        rssi = ap_info.rssi;
-    }
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiBeaconLostCount(uint32_t & beaconLostCount)
-{
-    beaconLostCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiCurrentMaxRate(uint64_t & currentMaxRate)
-{
-    currentMaxRate = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount)
-{
-    packetMulticastRxCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount)
-{
-    packetMulticastTxCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount)
-{
-    packetUnicastRxCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount)
-{
-    packetUnicastTxCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_GetWiFiOverrunCount(uint64_t & overrunCount)
-{
-    overrunCount = 0;
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR ConnectivityManagerImpl::_ResetWiFiNetworkDiagnosticsCounts()
-{
-    return CHIP_NO_ERROR;
-}
-
 } // namespace DeviceLayer
 } // namespace chip
