+++ src/platform/ESP32/DiagnosticDataProviderImpl.cpp
@@ -37,6 +37,60 @@
 #include "esp_system.h"
 #include "esp_wifi.h"
 
+using namespace ::chip;
+using namespace ::chip::TLV;
+using namespace ::chip::DeviceLayer;
+using namespace ::chip::DeviceLayer::Internal;
+using namespace ::chip::app::Clusters::GeneralDiagnostics;
+
+namespace {
+
+InterfaceType GetInterfaceType(const char * if_desc)
+{
+    if (strncmp(if_desc, "ap", strnlen(if_desc, 2)) == 0 || strncmp(if_desc, "sta", strnlen(if_desc, 3)) == 0)
+        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_WI_FI;
+    if (strncmp(if_desc, "openthread", strnlen(if_desc, 10)) == 0)
+        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_THREAD;
+    if (strncmp(if_desc, "eth", strnlen(if_desc, 3)) == 0)
+        return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_ETHERNET;
+    return InterfaceType::EMBER_ZCL_INTERFACE_TYPE_UNSPECIFIED;
+}
+
+#if CHIP_DEVICE_CONFIG_ENABLE_WIFI
+uint8_t MapAuthModeToSecurityType(wifi_auth_mode_t authmode)
+{
+    switch (authmode)
+    {
+    case WIFI_AUTH_OPEN:
+        return 1;
+    case WIFI_AUTH_WEP:
+        return 2;
+    case WIFI_AUTH_WPA_PSK:
+        return 3;
+    case WIFI_AUTH_WPA2_PSK:
+        return 4;
+    case WIFI_AUTH_WPA3_PSK:
+        return 5;
+    default:
+        return 0;
+    }
+}
+
+uint8_t GetWiFiVersionFromAPRecord(wifi_ap_record_t ap_info)
+{
+    if (ap_info.phy_11n)
+        return 3;
+    else if (ap_info.phy_11g)
+        return 2;
+    else if (ap_info.phy_11b)
+        return 1;
+    else
+        return 0;
+}
+#endif // CHIP_DEVICE_CONFIG_ENABLE_WIFI
+
+} // namespace
+
 namespace chip {
 namespace DeviceLayer {
 
@@ -139,6 +193,172 @@ CHIP_ERROR DiagnosticDataProviderImpl::GetBootReason(uint8_t & bootReason)
     }
     return CHIP_NO_ERROR;
 }
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetNetworkInterfaces(NetworkInterface ** netifpp)
+{
+    esp_netif_t * netif     = esp_netif_next(NULL);
+    NetworkInterface * head = NULL;
+    if (netif == NULL)
+    {
+        ChipLogError(DeviceLayer, "Failed to get network interfaces");
+    }
+    else
+    {
+        for (esp_netif_t * ifa = netif; ifa != NULL; ifa = esp_netif_next(ifa))
+        {
+            NetworkInterface * ifp = new NetworkInterface();
+            strncpy(ifp->Name, esp_netif_get_ifkey(ifa), Inet::InterfaceId::kMaxIfNameLength);
+            ifp->Name[Inet::InterfaceId::kMaxIfNameLength - 1] = '\0';
+            ifp->name                                          = CharSpan(ifp->Name, strlen(ifp->Name));
+            ifp->fabricConnected                               = true;
+            ifp->type                                          = GetInterfaceType(esp_netif_get_desc(ifa));
+            ifp->offPremiseServicesReachableIPv4               = false;
+            ifp->offPremiseServicesReachableIPv6               = false;
+            if (esp_netif_get_mac(ifa, ifp->MacAddress) != ESP_OK)
+            {
+                ChipLogError(DeviceLayer, "Failed to get network hardware address");
+            }
+            else
+            {
+                ifp->hardwareAddress = ByteSpan(ifp->MacAddress, 6);
+            }
+            ifp->Next = head;
+            head      = ifp;
+        }
+    }
+    *netifpp = head;
+    return CHIP_NO_ERROR;
+}
+
+void DiagnosticDataProviderImpl::ReleaseNetworkInterfaces(NetworkInterface * netifp)
+{
+    while (netifp)
+    {
+        NetworkInterface * del = netifp;
+        netifp                 = netifp->Next;
+        delete del;
+    }
+}
+
+#if CHIP_DEVICE_CONFIG_ENABLE_WIFI
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiBssId(ByteSpan & BssId)
+{
+    wifi_ap_record_t ap_info;
+    esp_err_t err;
+    uint8_t macAddress[kMaxHardwareAddrSize];
+
+    err = esp_wifi_sta_get_ap_info(&ap_info);
+    if (err == ESP_OK)
+    {
+        memcpy(macAddress, ap_info.bssid, 6);
+    }
+    BssId = ByteSpan(macAddress, 6);
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiSecurityType(uint8_t & securityType)
+{
+    securityType = 0;
+    wifi_ap_record_t ap_info;
+    esp_err_t err;
+
+    err = esp_wifi_sta_get_ap_info(&ap_info);
+    if (err == ESP_OK)
+    {
+        securityType = MapAuthModeToSecurityType(ap_info.authmode);
+    }
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiVersion(uint8_t & wifiVersion)
+{
+    wifiVersion = 0;
+    wifi_ap_record_t ap_info;
+    esp_err_t err;
+    err = esp_wifi_sta_get_ap_info(&ap_info);
+    if (err == ESP_OK)
+    {
+        wifiVersion = GetWiFiVersionFromAPRecord(ap_info);
+    }
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiChannelNumber(uint16_t & channelNumber)
+{
+    channelNumber = 0;
+    wifi_ap_record_t ap_info;
+    esp_err_t err;
+
+    err = esp_wifi_sta_get_ap_info(&ap_info);
+    if (err == ESP_OK)
+    {
+        channelNumber = ap_info.primary;
+    }
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiRssi(int8_t & rssi)
+{
+    rssi = 0;
+    wifi_ap_record_t ap_info;
+    esp_err_t err;
+
+    err = esp_wifi_sta_get_ap_info(&ap_info);
+
+    if (err == ESP_OK)
+    {
+        rssi = ap_info.rssi;
+    }
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiBeaconLostCount(uint32_t & beaconLostCount)
+{
+    beaconLostCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiCurrentMaxRate(uint64_t & currentMaxRate)
+{
+    currentMaxRate = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount)
+{
+    packetMulticastRxCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount)
+{
+    packetMulticastTxCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount)
+{
+    packetUnicastRxCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount)
+{
+    packetUnicastTxCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::GetWiFiOverrunCount(uint64_t & overrunCount)
+{
+    overrunCount = 0;
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR DiagnosticDataProviderImpl::ResetWiFiNetworkDiagnosticsCounts()
+{
+    return CHIP_NO_ERROR;
+}
+#endif // CHIP_DEVICE_CONFIG_ENABLE_WIFI
 
 } // namespace DeviceLayer
 } // namespace chip
