+++ src/platform/Linux/ConnectivityManagerImpl.h
@@ -110,10 +110,15 @@ public:
     CHIP_ERROR ProvisionWiFiNetwork(const char * ssid, const char * key);
     void StartWiFiManagement();
     bool IsWiFiManagementStarted();
+    CHIP_ERROR GetWiFiBssId(ByteSpan & value);
+    CHIP_ERROR GetWiFiSecurityType(uint8_t & securityType);
+    CHIP_ERROR GetWiFiVersion(uint8_t & wiFiVersion);
 #endif
 
+    const char * GetEthernetIfName() { return (mEthIfName[0] == '\0') ? nullptr : mEthIfName; }
+
 #if CHIP_DEVICE_CONFIG_ENABLE_WIFI
-    static const char * GetWiFiIfName() { return sWiFiIfName; }
+    static const char * GetWiFiIfName() { return (sWiFiIfName[0] == '\0') ? nullptr : sWiFiIfName; }
 #endif
 
 private:
@@ -136,9 +141,6 @@ private:
 
     WiFiAPMode _GetWiFiAPMode();
     CHIP_ERROR _SetWiFiAPMode(WiFiAPMode val);
-    CHIP_ERROR _GetWiFiBssId(ByteSpan & value);
-    CHIP_ERROR _GetWiFiSecurityType(uint8_t & securityType);
-    CHIP_ERROR _GetWiFiVersion(uint8_t & wiFiVersion);
     bool _IsWiFiAPActive();
     bool _IsWiFiAPApplicationControlled();
     void _DemandStartWiFiAP();
@@ -160,38 +162,7 @@ private:
     static std::mutex mWpaSupplicantMutex;
 #endif
 
-    void _ReleaseNetworkInterfaces(NetworkInterface * netifp);
-    CHIP_ERROR _GetNetworkInterfaces(NetworkInterface ** netifpp);
-    CHIP_ERROR _GetEthPHYRate(uint8_t & pHYRate);
-    CHIP_ERROR _GetEthFullDuplex(bool & fullDuplex);
-    CHIP_ERROR _GetEthTimeSinceReset(uint64_t & timeSinceReset);
-    CHIP_ERROR _GetEthPacketRxCount(uint64_t & packetRxCount);
-    CHIP_ERROR _GetEthPacketTxCount(uint64_t & packetTxCount);
-    CHIP_ERROR _GetEthTxErrCount(uint64_t & txErrCount);
-    CHIP_ERROR _GetEthCollisionCount(uint64_t & collisionCount);
-    CHIP_ERROR _GetEthOverrunCount(uint64_t & overrunCount);
-    CHIP_ERROR _ResetEthNetworkDiagnosticsCounts();
-
-#if CHIP_DEVICE_CONFIG_ENABLE_WIFI
-    CHIP_ERROR _GetWiFiChannelNumber(uint16_t & channelNumber);
-    CHIP_ERROR _GetWiFiRssi(int8_t & rssi);
-    CHIP_ERROR _GetWiFiBeaconLostCount(uint32_t & beaconLostCount);
-    CHIP_ERROR _GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount);
-    CHIP_ERROR _GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount);
-    CHIP_ERROR _GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount);
-    CHIP_ERROR _GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount);
-    CHIP_ERROR _GetWiFiCurrentMaxRate(uint64_t & currentMaxRate);
-    CHIP_ERROR _GetWiFiOverrunCount(uint64_t & overrunCount);
-    CHIP_ERROR _ResetWiFiNetworkDiagnosticsCounts();
-#endif
-
     // ==================== ConnectivityManager Private Methods ====================
-
-    CHIP_ERROR ResetEthernetStatsCount();
-
-#if CHIP_DEVICE_CONFIG_ENABLE_WIFI
-    CHIP_ERROR ResetWiFiStatsCount();
-#endif
 
 #if CHIP_DEVICE_CONFIG_ENABLE_WPA
     void DriveAPState();
@@ -209,11 +180,6 @@ private:
 
     // ===== Private members reserved for use by this class only.
 
-    uint64_t mEthPacketRxCount  = 0;
-    uint64_t mEthPacketTxCount  = 0;
-    uint64_t mEthTxErrCount     = 0;
-    uint64_t mEthCollisionCount = 0;
-    uint64_t mEthOverrunCount   = 0;
     char mEthIfName[IFNAMSIZ];
 
 #if CHIP_DEVICE_CONFIG_ENABLE_WPA
@@ -223,16 +189,9 @@ private:
     System::Clock::Timestamp mLastAPDemandTime;
     System::Clock::Timeout mWiFiStationReconnectInterval;
     System::Clock::Timeout mWiFiAPIdleTimeout;
-    uint8_t mWiFiMacAddress[kMaxHardwareAddrSize];
 #endif
 
 #if CHIP_DEVICE_CONFIG_ENABLE_WIFI
-    uint32_t mBeaconLostCount        = 0;
-    uint32_t mPacketMulticastRxCount = 0;
-    uint32_t mPacketMulticastTxCount = 0;
-    uint32_t mPacketUnicastRxCount   = 0;
-    uint32_t mPacketUnicastTxCount   = 0;
-    uint64_t mOverrunCount           = 0;
     static char sWiFiIfName[IFNAMSIZ];
 #endif
 };
